BookAppointmentScreen update redux internally after booking appointment, no longer needs to wait loading from firebase
BookAppointmentScreen now wait for the data retrieval from server to complete before the user can proceed to book an appointment to make sure the newest restriction is applied to the time slot.

When videoCall ends for the patient side, the appointment is now automatically updated to completed without needing to load from server.

When videoCall ends for the patient side, the UI will return to calendar screen instead of AppointmentDetailScreen